[Fix] firebase auth에서 쓰이던 uid로 선언된 부분들 및 명칭을 id/userId로 수정 #34

+++ src/services/element-service.ts
@@ -7,13 +7,13 @@ import { PostgrestSingleResponse } from '@supabase/supabase-js';
 import { normalizeError } from '../utils/util-convert';
 
 export default class ElementService {
-  static async getElementsData(userUid: string, id: string) {
+  static async getElementsData(userId: string, id: string) {
     try {
       const elementsTable = await supabaseDatabase('elements');
       const { data: elementsData }: PostgrestSingleResponse<ElementsBasicType> = await elementsTable
         .select('*')
         .eq('id', id)
-        .eq('userUid', userUid)
+        .eq('userId', userId)
         .single();
 
       return elementsData;
@@ -23,12 +23,12 @@ export default class ElementService {
     }
   }
 
-  static async postElementsData(userUid: string, id: string, data: ElementsBasicType) {
+  static async postElementsData(userId: string, id: string, data: ElementsBasicType) {
     try {
       const userInfo = localStorage.getItem('userInfo') ?? '';
       const parseUserInfo = getParsedJsonData<UserInfoType>(userInfo);
       const { error: transactionError } = await supabase.rpc('post_elements_data', {
-        userUid,
+        userId,
         id,
         data: {
           ...data,
@@ -44,15 +44,15 @@ export default class ElementService {
         throw transactionError;
       }
 
-      // elements > userUid > id > data // 더 쪼개야하나?
+      // elements > userId > id > data // 더 쪼개야하나?
       //?CONCERN: postgreSQL로 넘어오면서 향후 어떤식으로 데이터 관리해야할지 조금 고민...
       const elementsTable = await supabaseDatabase('elements');
       const travelsTable = await supabaseDatabase('travels');
       const usersTable = await supabaseDatabase('users');
 
-      const { error: elementsError } = await elementsTable.insert({ id, userUid, ...data });
+      const { error: elementsError } = await elementsTable.insert({ id, userId, ...data });
       const { error: travelsError } = await travelsTable.insert({
-        userUid,
+        userId,
         ...data.info,
         id,
       });
@@ -71,10 +71,10 @@ export default class ElementService {
     }
   }
 
-  static async deleteElementsData(userUid: string, id: string) {
+  static async deleteElementsData(userId: string, id: string) {
     try {
       const { error: transactionError } = await supabase.rpc('delete_elements_data', {
-        userUid,
+        userId,
         id,
       });
 
@@ -88,11 +88,8 @@ export default class ElementService {
       const { error: elementsError } = await elementsTable
         .delete()
         .eq('id', id)
-        .eq('userUid', userUid);
-      const { error: travelsError } = await travelsTable
-        .delete()
-        .eq('id', id)
-        .eq('userUid', userUid);
+        .eq('userId', userId);
+      const { error: travelsError } = await travelsTable.delete().eq('id', id).eq('userId', userId);
 
       if (elementsError || travelsError) {
         throw new Error(elementsError?.message || travelsError?.message);
